// packages/amplify-util-uibuilder/src/local_modules/aws-sdk/clients/amplifyuibuilder.d.ts
  export type AssociatedFieldsList = String[];
  export type Boolean = boolean;
  export interface CodegenFeatureFlags {
    isRelationshipSupported?: Boolean;
    isNonModelSupported?: Boolean;
  }
  export interface CodegenGenericDataEnum {
    values: CodegenGenericDataEnumValuesList;
  }
  export type CodegenGenericDataEnumValuesList = String[];
  export type CodegenGenericDataEnums = {[key: string]: CodegenGenericDataEnum};
  export interface CodegenGenericDataField {
    dataType: CodegenGenericDataFieldDataType;
    dataTypeValue: String;
    required: Boolean;
    readOnly: Boolean;
    isArray: Boolean;
    relationship?: CodegenGenericDataRelationshipType;
  }
  export type CodegenGenericDataFieldDataType = "ID"|"String"|"Int"|"Float"|"AWSDate"|"AWSTime"|"AWSDateTime"|"AWSTimestamp"|"AWSEmail"|"AWSURL"|"AWSIPAddress"|"Boolean"|"AWSJSON"|"AWSPhone"|"Enum"|"Model"|"NonModel"|string;
  export type CodegenGenericDataFields = {[key: string]: CodegenGenericDataField};
  export interface CodegenGenericDataModel {
    fields: CodegenGenericDataFields;
    isJoinTable?: Boolean;
    primaryKeys: CodegenPrimaryKeysList;
  }
  export type CodegenGenericDataModels = {[key: string]: CodegenGenericDataModel};
  export interface CodegenGenericDataNonModel {
    fields: CodegenGenericDataNonModelFields;
  }
  export type CodegenGenericDataNonModelFields = {[key: string]: CodegenGenericDataField};
  export type CodegenGenericDataNonModels = {[key: string]: CodegenGenericDataNonModel};
  export interface CodegenGenericDataRelationshipType {
    type: GenericDataRelationshipType;
    relatedModelName: String;
    relatedModelFields?: RelatedModelFieldsList;
    canUnlinkAssociatedModel?: Boolean;
    relatedJoinFieldName?: String;
    relatedJoinTableName?: String;
    belongsToFieldOnRelatedModel?: String;
    associatedFields?: AssociatedFieldsList;
    isHasManyIndex?: Boolean;
  }
  export interface CodegenJob {
    id: Uuid;
    appId: String;
    environmentName: String;
    status?: CodegenJobStatus;
    statusMessage?: String;
    asset?: CodegenJobAsset;
    tags?: Tags;
    createdAt?: SyntheticTimestamp_date_time;
    modifiedAt?: SyntheticTimestamp_date_time;
  }
  export interface CodegenJobAsset {
    downloadUrl?: String;
    expiresAt?: SyntheticTimestamp_date_time;
  }
  export interface CodegenJobGenericDataSchema {
    dataSourceType: CodegenJobGenericDataSourceType;
    models: CodegenGenericDataModels;
    enums: CodegenGenericDataEnums;
    nonModels: CodegenGenericDataNonModels;
  }
  export type CodegenJobGenericDataSourceType = "DataStore"|string;
  export interface CodegenJobRenderConfig {
    react?: ReactStartCodegenJobData;
  }
  export type CodegenJobStatus = "in_progress"|"failed"|"succeeded"|string;
  export interface CodegenJobSummary {
    appId: String;
    environmentName: String;
    id: Uuid;
    createdAt?: SyntheticTimestamp_date_time;
    modifiedAt?: SyntheticTimestamp_date_time;
  }
  export type CodegenJobSummaryList = CodegenJobSummary[];
  export type CodegenPrimaryKeysList = String[];
// ...
  export type Float = number;
// ...
  export type GenericDataRelationshipType = "HAS_MANY"|"HAS_ONE"|"BELONGS_TO"|string;
// ...
  export type RelatedModelFieldsList = String[];
// ...
  export type String = string;
  export type SyntheticTimestamp_date_time = Date;
  export type TagKey = string;
  export type TagValue = string;
  export type Tags = {[key: string]: TagValue};
// ...
  export type Uuid = string;